Remove the FeatureFlag defendants_search, and the logic related to CDA V1, which is unused.
