+++ spec/features/hearing_sorting_v2_spec.rb
@@ -8,7 +8,6 @@ RSpec.feature 'Hearing sorting', :vcr, type: :feature do
 
   before do
     allow(FeatureFlag).to receive(:enabled?).with(:hearing_summaries).and_return(true)
-    allow(FeatureFlag).to receive(:enabled?).with(:defendants_search).and_return(true)
     sign_in user
     visit "prosecution_cases/#{case_urn}"
   end
