+++ spec/features/search/cd_api/defendant_by_reference_spec.rb
@@ -4,7 +4,6 @@ RSpec.feature 'Defendant by reference search', :vcr, :js, type: :feature do
   let(:user) { create(:user) }
 
   before do
-    allow(FeatureFlag).to receive(:enabled?).with(:defendants_search).and_return(true)
     sign_in user
   end
