--- spec/fixtures/vcr_cassettes/spec/features/search/defendant_by_reference_spec.yml
@@ -1,107 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: http://localhost:9292/api/internal/v1/prosecution_cases?filter%5Bnational_insurance_number%5D=GP181930B&include=defendants,defendants.offences
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v1.10.0
-      Content-Type:
-      - application/vnd.api+json
-      Accept:
-      - application/vnd.api+json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-      Accept-Encoding:
-      - gzip,deflate
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      Content-Type:
-      - application/vnd.api+json; charset=utf-8
-      Etag:
-      - W/"af7377bfb0250d886b30509dc81bcbd7"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      X-Request-Id:
-      - 5fe4eb14-d4f7-40a0-9d91-752a10524573
-      X-Runtime:
-      - '0.049336'
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: UTF-8
-      string: '{"data":[{"id":"61ae64af-e4b7-4476-928f-5b9bf8078ecb","type":"prosecution_cases","attributes":{"prosecution_case_reference":"CXBNSYAVIL"},"relationships":{"defendants":{"data":[{"id":"714fe7be-819e-4a10-97af-46922400513f","type":"defendants"}]}}}],"included":[{"id":"714fe7be-819e-4a10-97af-46922400513f","type":"defendants","attributes":{"name":"Maria
-        Stoltenberg Prosacco","date_of_birth":"1962-02-03","national_insurance_number":"GP181930B","arrest_summons_number":"8KMZF92CXNEH","maat_reference":null},"relationships":{"offences":{"data":[{"id":"f394c472-8bad-4017-af20-03d206b8e876","type":"offences"}]}}},{"id":"f394c472-8bad-4017-af20-03d206b8e876","type":"offences","attributes":{"code":"Random
-        string","order_index":1,"title":"Random string","mode_of_trial":"Random string"}}]}'
-    http_version: null
-  recorded_at: Mon, 20 Apr 2020 13:45:54 GMT
-- request:
-    method: get
-    uri: http://localhost:9292/api/internal/v1/prosecution_cases?filter%5Bnational_insurance_number%5D=GP999999B&include=defendants,defendants.offences
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v1.10.0
-      Content-Type:
-      - application/vnd.api+json
-      Accept:
-      - application/vnd.api+json
-      Authorization:
-      - Bearer <BEARER_TOKEN>
-      Accept-Encoding:
-      - gzip,deflate
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      X-Frame-Options:
-      - SAMEORIGIN
-      X-Xss-Protection:
-      - 1; mode=block
-      X-Content-Type-Options:
-      - nosniff
-      X-Download-Options:
-      - noopen
-      X-Permitted-Cross-Domain-Policies:
-      - none
-      Referrer-Policy:
-      - strict-origin-when-cross-origin
-      Content-Type:
-      - application/vnd.api+json; charset=utf-8
-      Etag:
-      - W/"ba5f3ea40e95f49bce11942f375ebd38"
-      Cache-Control:
-      - max-age=0, private, must-revalidate
-      X-Request-Id:
-      - 1360ad17-ab5d-4c39-bdce-20841b30bb6c
-      X-Runtime:
-      - '0.016007'
-      Transfer-Encoding:
-      - chunked
-    body:
-      encoding: UTF-8
-      string: '{"data":null}'
-    http_version: null
-  recorded_at: Mon, 20 Apr 2020 13:45:57 GMT
-recorded_with: VCR 5.1.0
