+++ spec/requests/search/defendant_by_name_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe 'Defendant by name and dob search', type: :request do
     expect(response).to render_template('searches/new')
   end
 
-  context 'when posting a query', :vcr do
+  context 'when posting a query', :stub_defendants_name_search do
     let(:search_params) do
       {
         search:
@@ -41,7 +41,7 @@ RSpec.describe 'Defendant by name and dob search', type: :request do
 
       it 'assigns array of results' do
         expect(assigns(:results))
-          .to include(an_instance_of(CourtDataAdaptor::Resource::Defendant))
+          .to include(an_instance_of(CdApi::Defendant))
       end
 
       it 'renders searches/_results' do
@@ -57,7 +57,7 @@ RSpec.describe 'Defendant by name and dob search', type: :request do
       end
     end
 
-    context 'when no results', :stub_no_results, type: :request do
+    context 'when no results', :stub_no_v2_results, type: :request do
       before do
         allow_any_instance_of(Search).to receive(:execute).and_return([])
         post '/searches', params: search_params
